cartpole-balancing/requirements.txt
gym==0.26.2
gymnasium==0.29.1
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
matplotlib==3.7.2
mlflow==2.7.1
flask==2.3.3
pandas==2.0.3
seaborn==0.12.2
python-dotenv==1.0.0
scikit-learn==1.3.0